fix quit stretch bug
When pressing the Quit button while in a game, the left panel would
stretch super tall while animating out, then return to normal when the
transition animation was complete. This was a two-fold issue.

The first issue is that the `ResizeObserver` in the `Renderer` was
continuing to resize the canvas when animating out, causing lots of
weirdness in the height of the UI. I solved this by hooking the
`Renderer` up to the global store and preventing resizing if `isRunning`
has been flipped, then forcing the `GameManager` to stop sooner (when
the Quit button is clicked).

The second issue was that the `main` element was losing the `play` class
when animating out. This caused the canvas to lose important styles,
causing further weirdness with UI sizing. I fixed this by abstracting
the necessary styles into more universal components.