fix stale action caches
Using `keyState` as the action cache key didn't work because it would be
undefined when the key was released, preventing the stale cache from
being deleted. It also didn't make sense, because we don't want multiple
caches for a single action. We're now using the action label as the
cache key.